completion: test & ignore generalized var
